Refactor Supabase client import paths in ConversionResults and Dashboard components
- Updated import statements for the Supabase client to use the new path from '@/integrations/supabase/client'.
- Ensured consistency in the integration approach across components.
